requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
dj-stripe==2.4.1
Django==3.1.5
django-cors-headers==3.6.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
jsonfield==3.1.0
psycopg2==2.8.6
PyJWT==2.0.0
pytz==2020.5
requests==2.25.1
selenium==3.141.0
sqlparse==0.4.1
stripe==2.55.1
undetected-chromedriver==1.5.2
urllib3==1.26.2
whitenoise==5.2.0
braintree>=4.5.0
rq>=1.7.0
redis>=3.5.3
django-braces>=1.14.0
